requirements.txt
PySide2>=5.15.0

QDarkStyle~=2.8.1
PyYAML~=5.3.1
numpy>=1.19.0
perfplot>=0.8.1
pandas>=1.0.5
methodtools>=0.4.2
qimage2ndarray>=1.8.3
notifyr>=1.1
pytest>=6.0.1
setuptools>=49.2.0
Cython>=0.29.21
multipledispatch>=0.6.0
